se ha ha mejorado las entitys como su id, y los repos

+++ src/main/java/pe/sia/persistence/repository/activosRepository/ActivoInformaticoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.sia.persistence.entity.activos.ActivoInformatico;
 
 @Repository
-public interface ActivoInformaticoRepository extends JpaRepository<ActivoInformatico, Long> {
+public interface ActivoInformaticoRepository extends JpaRepository<ActivoInformatico, Integer> {
 
 }
