+++ src/main/java/pe/sia/service/interfaces/ProveedorService.java
@@ -7,9 +7,9 @@ import pe.sia.presentation.dto.activosDTO.ProveedorDTO;
 public interface ProveedorService {
 
     ProveedorDTO createProveedor(ProveedorDTO proveedorDTO);
-    ProveedorDTO updateProveedor(Long idProveedor, ProveedorDTO proveedorDTO);
-    void deleteProveedor(Long idProveedor);
-    ProveedorDTO getProveedorById(Long idProveedor);
+    ProveedorDTO updateProveedor(Integer idProveedor, ProveedorDTO proveedorDTO);
+    void deleteProveedor(Integer idProveedor);
+    ProveedorDTO getProveedorById(Integer idProveedor);
     List<ProveedorDTO> getAllProveedores();
 
 }
